requirements.txt
# Dev environment
pip
autopep8

# App
flask
pymongo
python-dotenv